LikeOptions extracted out
