+++ src/components/PostItem/PostItem.tsx
@@ -1,6 +1,6 @@
-import React, {useEffect, useRef, useState} from 'react';
+import React, {useRef, useState} from 'react';
 import {Animated, Image, TouchableOpacity, View} from 'react-native';
-import {TouchableWithoutFeedback} from 'react-native';
+import {Share} from 'react-native';
 import {
   faBookmark,
   faCircleXmark,
@@ -8,9 +8,6 @@ import {
   faEllipsis,
   faEyeSlash,
   faFlag,
-  faHeart,
-  faLaughSquint,
-  faLightbulb,
   faShareFromSquare,
   faThumbsUp,
 } from '@fortawesome/free-solid-svg-icons';
@@ -27,65 +24,10 @@ import {
   TUserInfo,
 } from '@DevEx/utils/types/types';
 
+import LikeOptions from '../LikeOptions/LikeOptions';
 import {Text} from '../Text/text';
 
 import createStyles from './PostItem.styles';
-
-const LikeOptions = ({
-  closeLiked,
-  setLiked,
-  animatedValues,
-}: {
-  closeLiked: (value: boolean) => void;
-  setLiked: (value: boolean) => void;
-  animatedValues: {
-    animatedPosition: any;
-    animatedOpacity: any;
-  };
-}) => {
-  const styles = useThemedStyles(createStyles);
-
-  useEffect(() => {
-    Animated.timing(animatedValues.animatedPosition, {
-      toValue: 40,
-      duration: 200,
-      useNativeDriver: false,
-    }).start();
-    Animated.timing(animatedValues.animatedOpacity, {
-      toValue: 1,
-      duration: 200,
-      useNativeDriver: false,
-    }).start();
-
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, []);
-
-  return (
-    <View style={styles.LikedOptionsPillContainer}>
-      <Animated.View
-        style={[
-          styles.LikedOptionsPillAnimatedContainer,
-          {
-            opacity: animatedValues.animatedOpacity,
-            bottom: animatedValues.animatedPosition,
-          },
-        ]}>
-        <FontAwesomeIcon icon={faThumbsUp} color={colors.blue} />
-        <FontAwesomeIcon icon={faHeart} color={colors.criticalRed} />
-        {/* {/* <FontAwesomeIcon icon={} /> */}
-        <FontAwesomeIcon
-          icon={faLaughSquint}
-          secondaryColor={colors.black}
-          color={colors.yellow}
-        />
-        <FontAwesomeIcon icon={faLightbulb} />
-      </Animated.View>
-      <TouchableWithoutFeedback onPressOut={() => closeLiked(false)}>
-        <Text textStyle={styles.LikedOptionsPillText} text=" Tap to cancel" />
-      </TouchableWithoutFeedback>
-    </View>
-  );
-};
 
 const PostItem = ({
   item,
@@ -120,7 +62,11 @@ const PostItem = ({
     {
       name: 'Share Via',
       icon: faShareFromSquare,
-      onPress: () => {},
+      onPress: async () =>
+        await Share.share({
+          title: item.data.content,
+          url: `www.DevNotion.com/test/post/${item.id}`,
+        }),
       color: colors.grey60,
       iconSize: 20,
     },
@@ -256,11 +202,11 @@ const PostItem = ({
               onLongPress={openLikedOptions}
               onPress={() => onPressLiked()}
               style={[
-                styles.PostInfoLikeButton,
                 // eslint-disable-next-line react-native/no-inline-styles
                 {
                   width: likedLength.toFixed().length >= 2 ? 37 : 30,
                 },
+                styles.PostInfoLikeButton,
               ]}>
               <FontAwesomeIcon
                 icon={faThumbsUp}
@@ -276,11 +222,11 @@ const PostItem = ({
                 })
               }
               style={[
-                styles.PostItemCommentContainer,
                 // eslint-disable-next-line react-native/no-inline-styles
                 {
                   width: item.comments.length.toFixed().length >= 2 ? 40 : 30,
                 },
+                styles.PostItemCommentContainer,
               ]}>
               <FontAwesomeIcon icon={faComment} color={colors.grey40} />
               <Text text={item.comments.length} />
